+++ src/forms/EffectsListDialog.hpp
@@ -1,0 +1,12 @@
+
+#pragma once
+
+#include "forms/PersistantDialog.hpp"
+
+class EffectsListDialog : public PersistantDialog {
+    Q_OBJECT
+
+public:
+    explicit EffectsListDialog(QWidget *parent = nullptr);
+
+};